autorecovery: initial implementation of a new save/recovery for database documents
For the moment, the document only stores and loads itself, as before.

In this course, the AutoRecovery service now delegates the decision whether or not a document
needs to be stored (to the "salvage location") to the document itself, using the new method
XDocumentRecovery::wasModifiedSinceLastSave, superseded the previous E_MODIFIED_SINCE_LAST_AUTOSAVE
flag.

XDocumentRecovery::wasModifiedSinceLastSave has been implemented for SfxBaseModel, yielding the same
behavior as before.

It has also been implemented for database documents, where the implementation is somewhat "sloppy",
in that it returns always <TRUE/> when any of the opened sub components of the document is modified.